+++ Cargo.toml
@@ -16,11 +16,14 @@ byteorder = "1.4"
 matches = "0.1"
 
 [dependencies]
-x86 = { version = "0.51", features = ["unstable"] }
 log = "0.4"
 custom_error = { version = "1.9", default-features = false, features = ["unstable"] }
 bit_field = "0.10.1"
 phf = { version = "0.10.0", default-features = false }
+
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+x86 = { version = "0.51", features = ["unstable"] }
+
 
 [build-dependencies]
 pciid-parser = "0.5.0"
