+++ src/iomem.rs
@@ -8,6 +8,7 @@ use core::ops::Index;
 use core::ptr::NonNull;
 
 use custom_error::custom_error;
+
 use x86::current::paging::{IOAddr, PAddr, VAddr};
 
 // custom error for the IOMemory
